# api/openapi.yaml
openapi: 3.1.0
info:
  title: (S)AGE API
  version: 11.15.0
  description: |-
    (Sovereign) Agent Governed Experience — network REST API for memory
    submission, validation, and query.

    This machine-readable contract intentionally excludes the same-machine
    human CEREBRUM control plane under `/v1/dashboard/**`. Those routes are
    loopback-only management APIs, not remotely callable network APIs. Their
    authoritative contract is the “App-v23 CEREBRUM access-control routes”
    section of `docs/reference/rest-api.md`.
  contact:
    name: Dhillon Andrew Kannabhiran
    email: dhillon@levelupctf.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security — Ed25519 signature authentication via headers
# ---------------------------------------------------------------------------
security:
  - ed25519Auth: []
components:
...
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >-
        Ed25519 authentication is conveyed through three headers sent together:
        **X-Agent-ID** (hex-encoded public key), **X-Signature** (Ed25519
        signature of the canonical request payload), and **X-Timestamp** (unix
        epoch seconds). All three must be present on authenticated endpoints.
        Current clients should also send **X-Nonce**. After app-v17 activation,
        consensus binds delegated REST proofs to the exact canonical request,
        enforces the timestamp against block time, and consumes each proof once.

  # -----------------------------------------------------------------------
  # Parameters
  # -----------------------------------------------------------------------
...
  schemas:
...
    MCPTokenIssueRequest:
      type: object
      required:
        - agent_id
      properties:
        name:
          type: string
          description: Human label, e.g. "chatgpt-laptop".
        agent_id:
          type: string
          description: 64-char hex Ed25519 pubkey to mint for.
    MCPTokenIssueResponse:
      type: object
      properties:
        id:
          type: string
          description: Public token ID used for revocation.
        name:
          type: string
        agent_id:
          type: string
        token:
          type: string
          description: Base64url(32 random bytes). Shown ONCE — never stored in plaintext.
        created_at:
          type: string
          format: date-time
        use_hint:
          type: string
    MCPTokenSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        agent_id:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        reason_code:
          type: string
          description: Stable sanitized machine-readable denial reason, when available.
        remedy:
          type: string
          description: Exact operator action that can make a later request valid.
        retryable:
          type: boolean
          description: Whether automatically repeating the same request is safe.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Missing or invalid authentication headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
paths:
...
  /v1/mcp/tokens:
    post:
      operationId: issueMCPToken
      summary: Issue a bearer token for MCP clients that cannot sign Ed25519 requests.
      description: >-
        Ed25519 auth required (admin use). Token plaintext shown ONCE — not
        stored, only SHA-256 digest persisted.
      tags:
        - MCP Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPTokenIssueRequest"
      responses:
        "201":
          description: Token issued. Save the token value — it is never shown again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPTokenIssueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: MCP tokens not supported on this backend.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    get:
      operationId: listMCPTokens
      summary: List issued tokens as summaries (no token values returned).
      tags:
        - MCP Tokens
      responses:
        "200":
          description: Token summaries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/MCPTokenSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: MCP tokens not supported on this backend.
  /v1/mcp/tokens/{id}:
    delete:
      operationId: revokeMCPToken
      summary: Revoke a token by ID. Idempotent.
      tags:
        - MCP Tokens
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Token UUID.
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # -- Pipeline -----------------------------------------------------------
...